# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/disciplines": {
      "get": {
        "description": "Lista disciplinas com paginação e filtros. Usuários ADMIN veem todas as disciplinas, outros usuários veem apenas as disciplinas associadas a eles.",
        "operationId": "DisciplinesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número da página",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Itens por página (máximo 100)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Termo de busca por nome da disciplina",
            "schema": {
              "example": "matemática",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de disciplinas retornada com sucesso",
            "schema": {
              "example": {
                "data": [
                  {
                    "id": 1,
                    "slug": "mathematics",
                    "name": "Matemática"
                  },
                  {
                    "id": 2,
                    "slug": "portuguese",
                    "name": "Português"
                  },
                  {
                    "id": 5,
                    "slug": "history",
                    "name": "História"
                  }
                ],
                "meta": {
                  "page": 1,
                  "pageSize": 10,
                  "total": 3,
                  "totalPages": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedDisciplinesResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido ou ausente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Listar disciplinas",
        "tags": [
          "disciplines"
        ]
      }
    },
    "/users": {
      "get": {
        "description": "Lista usuários com informações completas, perfil e disciplinas. Apenas usuários ADMIN podem acessar esta funcionalidade.",
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número da página",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Itens por página (máximo 100)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Termo de busca por nome ou email do usuário",
            "schema": {
              "example": "joão",
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filtrar por perfil do usuário",
            "schema": {
              "enum": [
                "ADMIN",
                "EDITOR",
                "REVIEWER",
                "USER"
              ],
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filtrar por status ativo/inativo",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso",
            "schema": {
              "example": {
                "data": [
                  {
                    "uuid": "550e8400-e29b-41d4-a716-446655440000",
                    "name": "João Silva",
                    "email": "joao@email.com",
                    "role": "USER",
                    "isActive": true,
                    "emailVerifiedAt": "2024-01-15T10:30:00Z",
                    "lastLoginAt": "2024-01-20T14:30:00Z",
                    "createdAt": "2024-01-10T09:00:00Z",
                    "updatedAt": "2024-01-20T14:30:00Z",
                    "disciplines": [
                      {
                        "id": 1,
                        "slug": "mathematics",
                        "name": "Matemática"
                      },
                      {
                        "id": 2,
                        "slug": "portuguese",
                        "name": "Português"
                      }
                    ]
                  }
                ],
                "meta": {
                  "page": 1,
                  "pageSize": 10,
                  "total": 1,
                  "totalPages": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUsersResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido ou ausente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Usuário sem permissão para listar usuários",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Access denied. Only ADMIN can list users.",
                    "timestamp": "2024-01-15T10:30:00Z",
                    "path": "/users"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Listar usuários",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{uuid}/deactivate": {
      "patch": {
        "description": "Inativa um usuário no sistema, impedindo seu login. Apenas usuários ADMIN podem executar esta ação. Não é possível inativar a própria conta.",
        "operationId": "UsersController_deactivateUser",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "description": "UUID do usuário a ser inativado",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário inativado com sucesso",
            "schema": {
              "example": {
                "message": "Usuário inativado com sucesso",
                "userUuid": "550e8400-e29b-41d4-a716-446655440000",
                "isActive": false
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatusResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido ou ausente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Usuário sem permissão ou tentando inativar própria conta",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Cannot deactivate your own account",
                    "timestamp": "2024-01-15T10:30:00Z",
                    "path": "/users/550e8400-e29b-41d4-a716-446655440000/deactivate"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Inativar usuário",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{uuid}/activate": {
      "patch": {
        "description": "Ativa um usuário inativo no sistema, permitindo seu login novamente. Apenas usuários ADMIN podem executar esta ação.",
        "operationId": "UsersController_activateUser",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "description": "UUID do usuário a ser ativado",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário ativado com sucesso",
            "schema": {
              "example": {
                "message": "Usuário ativado com sucesso",
                "userUuid": "550e8400-e29b-41d4-a716-446655440000",
                "isActive": true
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatusResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido ou ausente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Usuário sem permissão para ativar usuários",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Access denied. Only ADMIN can activate users.",
                    "timestamp": "2024-01-15T10:30:00Z",
                    "path": "/users/550e8400-e29b-41d4-a716-446655440000/activate"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Ativar usuário",
        "tags": [
          "users"
        ]
      }
    }
  },
  "info": {
    "title": "Validador de Questões API",
    "description": "API para validação e gerenciamento de questões educacionais",
    "version": "1.0",
    "contact": {
      "name": "Suporte",
      "url": "https://validador.com/contato",
      "email": "suporte@validador.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento"
    },
    {
      "url": "https://api.validador.com",
      "description": "Produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Token JWT para autenticação",
        "in": "header"
      }
    },
    "schemas": {
...
      "UnauthorizedErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "Código de status HTTP",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "Mensagem de erro de autorização",
            "example": "Token inválido ou expirado"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp do erro",
            "example": "2024-01-15T10:30:00Z"
          },
          "path": {
            "type": "string",
            "description": "Caminho da requisição",
            "example": "/api/auth/login"
          }
        },
        "required": [
          "statusCode",
          "message",
          "timestamp",
          "path"
        ]
      },
...
      "DisciplineResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID da disciplina",
            "example": 1
          },
          "slug": {
            "type": "string",
            "description": "Slug da disciplina",
            "example": "history"
          },
          "name": {
            "type": "string",
            "description": "Nome da disciplina",
            "example": "História"
          }
        },
        "required": [
          "id",
          "slug",
          "name"
        ]
      },
...
      "NotFoundErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "Código de status HTTP",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "Mensagem de recurso não encontrado",
            "example": "Usuário não encontrado"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp do erro",
            "example": "2024-01-15T10:30:00Z"
          },
          "path": {
            "type": "string",
            "description": "Caminho da requisição",
            "example": "/api/auth/login"
          }
        },
        "required": [
          "statusCode",
          "message",
          "timestamp",
          "path"
        ]
      },
...
      "PaginatedDisciplinesResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Lista de disciplinas",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DisciplineResponseDto"
            }
          },
          "meta": {
            "type": "object",
            "description": "Metadados da paginação",
            "example": {
              "page": 1,
              "pageSize": 10,
              "total": 25,
              "totalPages": 3
            }
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UserDisciplineDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID da disciplina",
            "example": 1
          },
          "slug": {
            "type": "string",
            "description": "Slug da disciplina",
            "example": "mathematics"
          },
          "name": {
            "type": "string",
            "description": "Nome da disciplina",
            "example": "Matemática"
          }
        },
        "required": [
          "id",
          "slug",
          "name"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "description": "UUID único do usuário",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "description": "Nome do usuário",
            "example": "João Silva"
          },
          "email": {
            "type": "string",
            "description": "Email do usuário",
            "example": "joao@email.com"
          },
          "role": {
            "type": "string",
            "description": "Perfil do usuário",
            "enum": [
              "ADMIN",
              "EDITOR",
              "REVIEWER",
              "USER"
            ],
            "example": "USER"
          },
          "isActive": {
            "type": "boolean",
            "description": "Status ativo/inativo do usuário",
            "example": true
          },
          "emailVerifiedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data de verificação do email",
            "example": "2024-01-15T10:30:00Z",
            "nullable": true
          },
          "lastLoginAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data do último login",
            "example": "2024-01-20T14:30:00Z",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data de criação da conta",
            "example": "2024-01-10T09:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data da última atualização",
            "example": "2024-01-20T14:30:00Z"
          },
          "disciplines": {
            "description": "Lista de disciplinas que o usuário tem acesso",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDisciplineDto"
            }
          }
        },
        "required": [
          "uuid",
          "name",
          "email",
          "role",
          "isActive",
          "emailVerifiedAt",
          "lastLoginAt",
          "createdAt",
          "updatedAt",
          "disciplines"
        ]
      },
      "PaginatedUsersResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Lista de usuários",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponseDto"
            }
          },
          "meta": {
            "type": "object",
            "description": "Metadados da paginação",
            "example": {
              "page": 1,
              "pageSize": 10,
              "total": 25,
              "totalPages": 3
            }
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UserStatusResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Mensagem de confirmação",
            "example": "Usuário inativado com sucesso"
          },
          "userUuid": {
            "type": "string",
            "description": "UUID do usuário afetado",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "isActive": {
            "type": "boolean",
            "description": "Novo status do usuário",
            "example": false
          }
        },
        "required": [
          "message",
          "userUuid",
          "isActive"
        ]
      }
    }
  }
}